pro7-main/requirements.txt
pandas
scikit-learn
nltk
pypdf
python-docx
streamlit
joblib